app/src/main/java/com/xavier_carpentier/go4lunch/data/response/autocomplete_response/MatchedSubstring.java
package com.xavier_carpentier.go4lunch.data.response.autocomplete_response;

import androidx.annotation.NonNull;

import java.io.Serializable;
import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

public class MatchedSubstring implements Serializable
{

    @SerializedName("length")
    @Expose
    private Integer length;
    @SerializedName("offset")
    @Expose
    private Integer offset;
    private final static long serialVersionUID = 7515840066180545538L;

    /**
     * No args constructor for use in serialization
     *
     */
    @SuppressWarnings("unused")
    public MatchedSubstring() {
    }

    @SuppressWarnings("unused")
    public MatchedSubstring(Integer length, Integer offset) {
        super();
        this.length = length;
        this.offset = offset;
    }

    public Integer getLength() {
        return length;
    }

    public void setLength(Integer length) {
        this.length = length;
    }

    @SuppressWarnings("unused")
    public Integer getOffset() {
        return offset;
    }

    @SuppressWarnings("unused")
    public void setOffset(Integer offset) {
        this.offset = offset;
    }

    @NonNull
    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        sb.append(MatchedSubstring.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
        sb.append("length");
        sb.append('=');
        sb.append(((this.length == null)?"<null>":this.length));
        sb.append(',');
        sb.append("offset");
        sb.append('=');
        sb.append(((this.offset == null)?"<null>":this.offset));
        sb.append(',');
        if (sb.charAt((sb.length()- 1)) == ',') {
            sb.setCharAt((sb.length()- 1), ']');
        } else {
            sb.append(']');
        }
        return sb.toString();
    }

}